requirements.txt
Flask==0.12.2
gunicorn==19.7.1
google-cloud-pubsub==0.30.1
google-api-core==0.1.3
google-cloud-storage==1.6.0
google-cloud-datastore==1.4.0